support r1-v and fix vlm-r1
