+++ paddlemix/examples/vlm_r1/data_config/rec.yaml
@@ -1,0 +1,5 @@
+datasets:
+     - json_path: /root/paddlejob/workspace/env_run/liaojincheng/backup/code/add_vlm_r1/PaddleMIX/data/refcoco/refcoco_train_new.json
+     - json_path: /root/paddlejob/workspace/env_run/liaojincheng/backup/code/add_vlm_r1/PaddleMIX/data/refcoco/refcocop_train_new.json
+     - json_path: /root/paddlejob/workspace/env_run/liaojincheng/backup/code/add_vlm_r1/PaddleMIX/data/refcoco/refcocog_train_new.json
+    # - json_path: /root/paddlejob/workspace/env_run/liaojincheng/backup/code/add_vlm_r1/PaddleMIX/data/refcoco/refcoco_train_debug.json
