+++ paddlemix/examples/vlm_r1/eval/test_r1-v.py
@@ -1,3 +1,4 @@
+
 import argparse
 import json
 import json
@@ -7,17 +8,17 @@ import re
 import sys
 from pprint import pprint
 
+from tqdm import tqdm
 import paddle
 from paddlenlp.generation import GenerationConfig
 from paddlenlp.utils.import_utils import import_module
-from tqdm import tqdm
 
 sys.path.append('paddlemix/examples/vlm_r1')
 from open_r1_mllm.utils.tokenizer import get_processor
 from open_r1_mllm.utils.constant import TEMPLATE_MAPPING, MODEL_MAPPING
 
 def parse_args():
-    parser = argparse.ArgumentParser(description="Run VLM-R1 evaluation.")
+    parser = argparse.ArgumentParser(description="Run R1-V evaluation.")
     parser.add_argument("--model_name", type=str,default="Qwen2.5-VL-3B-Instruct", required=True, help="Name of the model.")
     parser.add_argument("--model_path", type=str,default="Qwen/Qwen2.5-VL-3B-Instruct", required=True, help="Path to the model checkpoint.")
     parser.add_argument("--data_root", type=str, required=True, help="Root directory of the dataset.")
@@ -30,93 +31,66 @@ def parse_args():
     parser.add_argument("--seed", type=int, default=42, help="Seed for inference.")
     return parser.parse_args()
 
-def extract_bbox_answer(method,content):
-    bbox_match = False
-    if method == "baseline":
-        bbox_pattern = r'(-?\d*\.?\d+),\s*(-?\d*\.?\d+),\s*(-?\d*\.?\d+),\s*(-?\d*\.?\d+)'
-        bbox_match = re.search(bbox_pattern, content)
-    else:
-        answer_tag_pattern = "\s*<answer>(.*?)</answer>"
-        bbox_pattern = r"\[([-+]?\d*\.\d+|\d+),\s*([-+]?\d*\.\d+|\d+),\s*([-+]?\d*\.\d+|\d+),\s*([-+]?\d*\.\d+|\d+)\]"
-
-        content_answer_match = re.search(answer_tag_pattern, content, re.DOTALL)
-        if content_answer_match:
-            content_answer = content_answer_match.group(1).strip()
-            bbox_match = re.search(bbox_pattern, content_answer)
-    if bbox_match:
-        bbox = [
-            float(bbox_match.group(1)),
-            float(bbox_match.group(2)),
-            float(bbox_match.group(3)),
-            float(bbox_match.group(4)),
-        ]
-        x1, y1, x2, y2 = bbox
-        return bbox, False
-    return [0, 0, 0, 0], False
-
-
-def iou(box1, box2):
-    inter_x1 = max(box1[0], box2[0])
-    inter_y1 = max(box1[1], box2[1])
-    inter_x2 = min(box1[2] - 1, box2[2] - 1)
-    inter_y2 = min(box1[3] - 1, box2[3] - 1)
-    if inter_x1 < inter_x2 and inter_y1 < inter_y2:
-        inter = (inter_x2 - inter_x1 + 1) * (inter_y2 - inter_y1 + 1)
-    else:
-        inter = 0
-    union = (
-        (box1[2] - box1[0]) * (box1[3] - box1[1])
-        + (box2[2] - box2[0]) * (box2[3] - box2[1])
-        - inter
-    )
-    return float(inter) / union
 
 def get_test_template(method):
     # TODO: Support other prompt
     if method =="r1":
-        QUESTION_TEMPLATE = "{Question} First output the thinking process in <think> </think> tags and then output the final answer in <answer> </answer> tags. Output the final answer in JSON format."
+        QUESTION_TEMPLATE = "{Question} First output the thinking process in <think> </think> and final answer (number) in <answer> </answer> tags."
     else:
-        QUESTION_TEMPLATE = "Locate {Question}, output its bbox coordinates using JSON format."
+        QUESTION_TEMPLATE = "{Question} First output the thinking process in <think> </think> and final answer (number) in <answer> </answer> tags."
     
     return QUESTION_TEMPLATE
+
+def extract_number_answer(output_str):
+    # Try to find the number within <answer> tags, if can not find, return None
+    answer_pattern = r'<answer>\s*(\d+)\s*</answer>'
+    match = re.search(answer_pattern, output_str)
+    
+    if match:
+        return int(match.group(1))
+    return None
 
 def main(args):
     steps = args.steps
     print("Steps: ", steps)
     MODEL_PATH = args.model_path
     os.makedirs("logs",exist_ok=True)
-    OUTPUT_PATH = "./logs/rec_results_{DATASET}_{MODEL}_{METHOD}_{STEPS}.json"
+    OUTPUT_PATH = "./logs/r1-v_results_{DATASET}_{MODEL}_{METHOD}_{STEPS}.json"
     BSZ = args.batch_size
     DATA_ROOT = args.data_root
     TEST_DATASETS = args.test_datasets
     IMAGE_ROOT = args.image_root
     random.seed(args.seed)
-
     template_name = TEMPLATE_MAPPING[args.model_name]
     process_vision_info = import_module(f"paddlemix.processors.{template_name}_processing.process_vision_info")
-    
-    # register model
+
     model_cls = import_module(f"paddlemix.models.{MODEL_MAPPING[args.model_name]}")
     model = model_cls.from_pretrained(
         MODEL_PATH,
         dtype="bfloat16",
-        attn_implementation="flash_attention_2"
     )
-    processor, tokenizer = get_processor(args.model_name,MODEL_PATH)
 
+    processor, tokenizer = get_processor(args.model_name,MODEL_PATH)
     sample_num = args.sample_num
     for ds in TEST_DATASETS:
         print(f"Processing {ds}...")
-        ds_path = os.path.join(DATA_ROOT, f"{ds}.json")
-        data = json.load(open(ds_path, "r"))
-        random.shuffle(data)
+        ds_path = os.path.join(DATA_ROOT, f"{ds}.jsonl")
+        # data = json.load(open(ds_path, "r"))
+        with open(ds_path, 'r') as file:
+            data = [json.loads(line) for line in file]
+        # random.shuffle(data)
         QUESTION_TEMPLATE = get_test_template(args.method)
-        data = data[:sample_num]
+        # data = data[:sample_num]
         messages = []
         for x in data:
-            image_path = os.path.join(IMAGE_ROOT, x["image"])
-            question = x['normal_caption'] if args.method=="baseline" else x['problem'] 
+            if x["image_path"].startswith("./"):
+                x["image_path"] = x["image_path"][2:]
+            image_path = os.path.join(IMAGE_ROOT, x["image_path"])
+            # question = x['normal_caption'] if args.method=="baseline" else x['problem'] 
+            question = x['question']
+            # SYSTEM_PROMPT = "A conversation between User and Assistant. The user asks a question, and the Assistant solves it. The assistant first thinks about the reasoning process in the mind and then provides the user with the answer. The reasoning process and answer are enclosed within <think> </think> and <answer> </answer> tags, respectively, i.e., <think> reasoning process here </think><answer> answer here </answer>"
             message = [
+                # {"role": "system", "content": SYSTEM_PROMPT},
                 {
                     "role": "user",
                     "content": [
@@ -155,6 +129,7 @@ def main(args):
             )
             # TODO
             # padding side bug 
+            # import pdb;pdb.set_trace()
             inputs.update(text_inputs)
             generation_config = GenerationConfig(
                 use_cache=True,
@@ -182,26 +157,26 @@ def main(args):
         correct_number = 0
         for input_example, model_output in zip(data, all_outputs):
             original_output = model_output
-            ground_truth = input_example["solution"]
-            ground_truth_normalized = input_example["normalized_solution"]
-            model_answer, normalized = extract_bbox_answer(args.method,original_output)
-            correct = 0
-            if model_answer is not None:
-                if not normalized and iou(model_answer, ground_truth) > 0.5:
-                    correct = 1
-                elif normalized and iou(model_answer, ground_truth_normalized) > 0.5:
-                    correct = 1
-            correct_number += correct
+            ground_truth = input_example['ground_truth']
+            model_answer = extract_number_answer(original_output)
+            
+            # Create a result dictionary for this example
             result = {
-                "question": QUESTION_TEMPLATE.format(Question=x["problem"]),
-                "ground_truth": ground_truth,
-                "model_output": original_output,
-                "extracted_answer": model_answer,
-                "correct": correct,
+                'question': input_example,
+                'ground_truth': ground_truth,
+                'model_output': original_output,
+                'extracted_answer': model_answer
             }
             final_output.append(result)
+            
+            # Count correct answers
+            if model_answer is not None and model_answer == ground_truth:
+                correct_number += 1
+
+        # Calculate and print accuracy
         accuracy = correct_number / len(data) * 100
-        print(f"\nAccuracy of {ds}: {accuracy:.2f}%")
+        print(f"\nAccuracy: {accuracy:.2f}%")
+
         output_path = OUTPUT_PATH.format(DATASET=ds,MODEL=args.model_name.lower(), METHOD=args.method, STEPS=steps)
         with open(output_path, "w") as f:
             json.dump({"accuracy": accuracy, "results": final_output}, f, indent=2)
