+++ paddlemix/examples/vlm_r1/sripts/preprocess_refcoco.py
